@@ -1,21 +1,20 @@
 require 'rails_helper'
 
-RSpec.describe "posts/edit", type: :view do
+RSpec.describe 'posts/edit', type: :view do
   before(:each) do
     @post = assign(:post, Post.create!(
-      :title => "MyString",
-      :body => "MyText"
-    ))
+                            title: 'MyString',
+                            body: 'MyText'
+                          ))
   end
 
-  it "renders the edit post form" do
+  it 'renders the edit post form' do
     render
 
-    assert_select "form[action=?][method=?]", post_path(@post), "post" do
+    assert_select 'form[action=?][method=?]', post_path(@post), 'post' do
+      assert_select 'input#post_title[name=?]', 'post[title]'
 
-      assert_select "input#post_title[name=?]", "post[title]"
-
-      assert_select "textarea#post_body[name=?]", "post[body]"
+      assert_select 'textarea#post_body[name=?]', 'post[body]'
     end
   end
 end
