@@ -1,21 +1,20 @@
 require 'rails_helper'
 
-RSpec.describe "posts/new", type: :view do
+RSpec.describe 'posts/new', type: :view do
   before(:each) do
     assign(:post, Post.new(
-      :title => "MyString",
-      :body => "MyText"
-    ))
+                    title: 'MyString',
+                    body: 'MyText'
+                  ))
   end
 
-  it "renders new post form" do
+  it 'renders new post form' do
     render
 
-    assert_select "form[action=?][method=?]", posts_path, "post" do
+    assert_select 'form[action=?][method=?]', posts_path, 'post' do
+      assert_select 'input#post_title[name=?]', 'post[title]'
 
-      assert_select "input#post_title[name=?]", "post[title]"
-
-      assert_select "textarea#post_body[name=?]", "post[body]"
+      assert_select 'textarea#post_body[name=?]', 'post[body]'
     end
   end
 end
