@@ -1,14 +1,14 @@
 require 'rails_helper'
 
-RSpec.describe "posts/show", type: :view do
+RSpec.describe 'posts/show', type: :view do
   before(:each) do
     @post = assign(:post, Post.create!(
-      :title => "Title",
-      :body => "MyText"
-    ))
+                            title: 'Title',
+                            body: 'MyText'
+                          ))
   end
 
-  it "renders attributes in <p>" do
+  it 'renders attributes in <p>' do
     render
     expect(rendered).to match(/Title/)
     expect(rendered).to match(/MyText/)
